requirements.txt
Flask==2.0.2
ipython==7.31.0
pytest==6.2.5
Pillow==4.3